Move the `enableAudienceGroup()` action to the audiences store.
